[Silver II] Title: 연속합, Time: 272 ms, Memory: 22548 KB -BaekjoonHub

+++ 백준/Silver/1912. 연속합/연속합.java
@@ -18,9 +18,9 @@ public class Main {
     }
 
     static void solve(){
-        long value=arr[0], result = arr[0];
+        int value=arr[0], result = arr[0];
         for(int i=1; i<n; i++){
-            long cur = arr[i];
+            int cur = arr[i];
             if(value < 0)
                 value = cur;
             else
